Refactored the settings classes.
