+++ classes/admin/class-placeholders.php
@@ -21,7 +21,7 @@ class Placeholders {
 	 * Contructor
 	 */
 	public function __construct() {
-		add_action( 'cmb2_init', array( $this, 'register_taxonomy_fields' ), 1 );
+		add_action( 'lsx_bd_settings_section_archive', array( $this, 'register_placeholder_fields' ), 5, 2 );
 	}
 
 	/**
@@ -40,22 +40,18 @@ class Placeholders {
 	}
 
 	/**
-	 * Configure Taxonomy Featured Images
+	 * Enable Business Directory Search settings only if LSX Search plugin is enabled.
 	 *
+	 * @param object $cmb The CMB2() class.
+	 * @param string $position either top of bottom.
 	 * @return void
 	 */
-	public function register_taxonomy_fields() {
-		$cmb    = new_cmb2_box(
-			array(
-				'id'           => '_term_details_metabox',
-				'title'        => esc_html__( 'Featured Image', 'lsx-business-directory' ),
-				'object_types' => array( 'term' ),
-				'taxonomies'   => array( 'lsx-bd-industry', 'lsx-bd-region' ),
-			)
-		);
-		$fields = \lsx\business_directory\includes\get_featured_image_field( 'lsx_bd' );
-		foreach ( $fields as $field ) {
-			$cmb->add_field( $field );
+	public function register_placeholder_fields( $cmb, $position ) {
+		if ( 'top' === $position ) {
+			$fields = \lsx\business_directory\includes\get_banner_fields( 'archive' );
+			foreach ( $fields as $field ) {
+				$cmb->add_field( $field );
+			}
 		}
 	}
 }
