+++ classes/admin/class-single.php
@@ -1,0 +1,69 @@
+<?php
+namespace lsx\business_directory\classes\admin;
+
+/**
+ * Featured Images for the Business Directory Taxonomies.
+ *
+ * @package lsx-business-directory
+ */
+class Single {
+
+	/**
+	 * Holds class instance
+	 *
+	 * @since 1.0.0
+	 *
+	 * @var      object \lsx\business_directory\classes\admin\Single()
+	 */
+	protected static $instance = null;
+
+	/**
+	 * Contructor
+	 */
+	public function __construct() {
+		add_action( 'lsx_bd_settings_page', array( $this, 'configure_settings_single_fields' ), 1, 1 );
+	}
+
+	/**
+	 * Return an instance of this class.
+	 *
+	 * @since 1.0.0
+	 *
+	 * @return    object \lsx\business_directory\classes\admin\Single()    A single instance of this class.
+	 */
+	public static function get_instance() {
+		// If the single instance hasn't been set, set it now.
+		if ( null == self::$instance ) {
+			self::$instance = new self();
+		}
+		return self::$instance;
+	}
+
+	/**
+	 * Configure Business Directory custom fields for the Settings page Single section.
+	 *
+	 * @param object $cmb new_cmb2_box().
+	 * @return void
+	 */
+	public function configure_settings_single_fields( $cmb ) {
+		$cmb->add_field(
+			array(
+				'id'          => 'settings_single',
+				'type'        => 'title',
+				'name'        => __( 'Single', 'lsx-business-directory' ),
+				'default'     => __( 'Single', 'lsx-business-directory' ),
+				'description' => __( 'The settings for the single business directory view.', 'lsx-business-directory' ),
+			)
+		);
+
+		$cmb->add_field(
+			array(
+				'name'             => esc_html__( 'Enquiry Form', 'lsx-business-directory' ),
+				'id'               => 'single_enquiry_form',
+				'type'             => 'select',
+				'show_option_none' => 'Choose a Form',
+				'options'          => lsx_bd_get_available_forms(),
+			)
+		);
+	}
+}
