+++ classes/class-admin.php
@@ -25,17 +25,28 @@ class Admin {
 	public $banners;
 
 	/**
+	 * Holds the admin the post type singles.
+	 *
+	 * @var object \lsx\business_directory\classes\admin\Archives();
+	 */
+	public $archives;
+
+	/**
+	 * Holds the admin for the post type archives.
+	 *
+	 * @var object \lsx\business_directory\classes\admin\Single();
+	 */
+	public $single;
+
+	/**
 	 * Contructor
 	 */
 	public function __construct() {
 		$this->load_classes();
 		// Enqueue scripts for all admin pages.
 		add_action( 'admin_enqueue_scripts', array( $this, 'assets' ) );
-
 		// Configure Settings page.
 		add_action( 'cmb2_admin_init', array( $this, 'configure_settings_fields' ) );
-		add_action( 'lsx_bd_settings_page', array( $this, 'configure_settings_single_fields' ), 1, 1 );
-		add_action( 'lsx_bd_settings_page', array( $this, 'configure_settings_archive_fields' ), 2, 1 );
 	}
 
 	/**
@@ -62,7 +73,13 @@ class Admin {
 		$this->banners = admin\Banners::get_instance();
 
 		require_once LSX_BD_PATH . 'classes/admin/class-term-thumbnail.php';
-		$this->banners = admin\Term_Thumbnail::get_instance();
+		$this->term_thumbnail = admin\Term_Thumbnail::get_instance();
+
+		require_once LSX_BD_PATH . 'classes/admin/class-archives.php';
+		$this->archives = admin\Archives::get_instance();
+
+		require_once LSX_BD_PATH . 'classes/admin/class-single.php';
+		$this->single = admin\Single::get_instance();
 	}
 
 	/**
@@ -71,7 +88,6 @@ class Admin {
 	 * @return void
 	 */
 	public function assets() {
-		// wp_enqueue_media();
 		wp_enqueue_script( 'media-upload' );
 		wp_enqueue_script( 'thickbox' );
 		wp_enqueue_style( 'thickbox' );
@@ -97,65 +113,5 @@ class Admin {
 			)
 		);
 		do_action( 'lsx_bd_settings_page', $this->cmb );
-	}
-
-	/**
-	 * Configure Business Directory custom fields for the Settings page Single section.
-	 *
-	 * @param object $cmb new_cmb2_box().
-	 * @return void
-	 */
-	public function configure_settings_single_fields( $cmb ) {
-		$cmb->add_field(
-			array(
-				'id'          => 'settings_single',
-				'type'        => 'title',
-				'name'        => __( 'Single', 'lsx-business-directory' ),
-				'default'     => __( 'Single', 'lsx-business-directory' ),
-				'description' => __( 'The settings for the single business directory view.', 'lsx-business-directory' ),
-			)
-		);
-
-		$cmb->add_field(
-			array(
-				'name'             => esc_html__( 'Enquiry Form', 'lsx-business-directory' ),
-				'id'               => 'single_enquiry_form',
-				'type'             => 'select',
-				'show_option_none' => 'Choose a Form',
-				'options'          => lsx_bd_get_available_forms(),
-			)
-		);
-	}
-
-	/**
-	 * Configure Business Directory custom fields for the Settings page General section.
-	 *
-	 * @param object $cmb new_cmb2_box().
-	 * @return void
-	 */
-	public function configure_settings_archive_fields( $cmb ) {
-		$cmb->add_field(
-			array(
-				'id'          => 'settings_archive',
-				'type'        => 'title',
-				'name'        => __( 'Archive', 'lsx-business-directory' ),
-				'default'     => __( 'Archive', 'lsx-business-directory' ),
-				'description' => __( 'Business Directory post type archive settings.', 'lsx-business-directory' ),
-			)
-		);
-		do_action( 'lsx_bd_settings_section_archive', $this->cmb, 'top' );
-		$cmb->add_field(
-			array(
-				'name'             => esc_html__( 'Layout option', 'lsx-business-directory' ),
-				'id'               => 'archive_grid_list',
-				'type'             => 'radio',
-				'show_option_none' => false,
-				'options'          => array(
-					'grid' => esc_html__( 'Grid', 'lsx-business-directory' ),
-					'list' => esc_html__( 'List', 'lsx-business-directory' ),
-				),
-			)
-		);
-		do_action( 'lsx_bd_settings_section_archive', $this->cmb, 'bottom' );
 	}
 }
